requirements.txt
# Document Intelligence Pipeline — Classification
# Python 3.11+

# ── Core ML ───────────────────────────────────────────────────────
sentence-transformers>=3.0.0
torch>=2.1.0
numpy>=1.26.0

# ── OCR ───────────────────────────────────────────────────────────
paddlepaddle-gpu>=2.6.0    # ou paddlepaddle pour CPU
paddleocr>=2.7.0
opencv-python-headless>=4.9.0

# ── Document Parsing ──────────────────────────────────────────────
PyMuPDF>=1.24.0            # fitz — extraction PDF native
python-docx>=1.1.0         # extraction DOCX

# ── API & Validation ──────────────────────────────────────────────
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
aiofiles>=24.1.0
pydantic>=2.9.0
pydantic-settings>=2.5.0

# ── Tests ─────────────────────────────────────────────────────────
pytest>=8.0.0
pytest-asyncio>=0.24.0
httpx>=0.27.0              # pour les tests FastAPI
